+++ src/components/features/testAuth/constants/queryKeys.ts
@@ -1,5 +1,4 @@
 export const testAuthKeys = {
   all: ['testAuth'] as const,
   login: () => [...testAuthKeys.all, 'login'] as const,
-  reissue: () => [...testAuthKeys.all, 'reissue'] as const,
 };
