+++ src/components/features/testAuth/constants/url.ts
@@ -1,4 +1,3 @@
 export const TEST_AUTH_ENDPOINTS = {
   login: '/test-auth/login',
-  reissue: '/auth/reissue',
 } as const;
