+++ src/components/features/users/constants/queryKeys.ts
@@ -1,4 +1,5 @@
 export const userKeys = {
   all: ['users'] as const,
   detail: () => [...userKeys.all, 'detail'] as const,
+  reissue: () => [...userKeys.all, 'reissue'] as const,
 };
