+++ src/components/features/users/constants/url.ts
@@ -1,3 +1,4 @@
 export const USER_ENDPOINTS = {
   detail: 'users/me',
+  reissue: '/auth/reissue',
 } as const;
